[FIX] web_responsive: Fix blank screen on search (support menus w/o xmlid)
If a menu item is created manually via the Odoo UI, then it (likely) won't have an xmlid. As a result, the `t-foreach` fails, resulting in the whole Odoo UI crashing (a blank white screen is shown). This code should use `id` as the `t-key` instead, as that is *always* available, even for menu items created manually.